Move metrics from controlplane to sdk/monitor/metrics
