+++ sdk/monitor/crossconnect/crossconnect_monitor_server.go
@@ -4,8 +4,8 @@ import (
 	"github.com/golang/protobuf/ptypes/empty"
 
 	"github.com/networkservicemesh/networkservicemesh/controlplane/api/crossconnect"
-	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/metrics"
 	"github.com/networkservicemesh/networkservicemesh/sdk/monitor"
+	"github.com/networkservicemesh/networkservicemesh/sdk/monitor/metrics"
 )
 
 // MonitorServer is a monitor.Server for crossconnect GRPC API
